+++ libraries/KT0803/src/KT0803.cpp
@@ -1,7 +1,7 @@
 //
 //    FILE: KT0803.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 // PURPOSE: Arduino Library for KT0803 and KT0803K FM transmitter.
 //     URL: https://github.com/RobTillaart/KT0803
 
@@ -12,7 +12,7 @@
 //  REGISTERS on page 7 datasheet
 
 
-KT0803::KT0803(TwoWire * wire)  
+KT0803::KT0803(TwoWire * wire)
 {
   _address  = 0x3E;
   _wire = wire;
@@ -41,8 +41,8 @@ bool KT0803::isConnected()
 bool KT0803::setFrequency(float MHz)
 {
   if ((MHz < 70) || (MHz > 108)) return false;
-  //  steps 50 KHz
-  return setChannel(round(MHz * 20));  
+  //  steps 50 KHz although KT0803 will truncate to 100 KHz.
+  return setChannel(round(MHz * 20));
 }
 
 
@@ -60,12 +60,15 @@ bool KT0803::setChannel(uint16_t channel)
   //  need to split over 3 registers
   //  register 2 part skipped (always 0) for KT0803
   uint16_t ch = channel >> 1;
-  //  register 0
-  if (writeData(0x00, ch & 0xFF) == false) return false;
-  //  register 1
-  uint8_t data = readData(0x01) & 0xF8;  //  keep other bits
-  data |= ((ch >> 8) & 0x07);
-  return writeData(0x01, data);
+
+  uint8_t register0 = ch & 0xFF;  //  CHSEL[8:1]
+  if (writeData(0x00, register0) == false) return false;
+
+  ch >>= 8;
+  uint8_t register1 = readData(0x01);
+  register1 &= 0xF8;         //  CHSEL[11:9]
+  register1 |= (ch & 0x07);  //  CHSEL[11:9]
+  return writeData(0x01, register1);
 }
 
 
@@ -74,6 +77,7 @@ uint16_t KT0803::getChannel()
   uint16_t channel = readData(0x01) & 0x07;
   channel <<= 8;
   channel |= readData(0x00);
+  channel <<= 1;
   return channel;
 }
 
@@ -106,11 +110,11 @@ bool KT0803::setRFGain(uint8_t rfgain)
   data |= (rfgain & 0x03) << 6;
   writeData(0x01, data);
   // bit 2
-  data = readData(0x13) & 0x7F; 
+  data = readData(0x13) & 0x7F;
   data |= (rfgain & 0x04) << 5;
   writeData(0x13, data);
   // bit 3
-  data = readData(0x02) & 0xBF; 
+  data = readData(0x02) & 0xBF;
   data |= (rfgain & 0x08) << 3;
   writeData(0x02, data);
   return true;
@@ -206,7 +210,7 @@ int KT0803::readData(uint8_t reg)
 {
   _wire->beginTransmission(_address);
   _wire->write(reg);
-  _wire->endTransmission();
+  _wire->endTransmission(false);  //  explicit no STOP  fig 3 page 4
 
   if (_wire->requestFrom(_address, (uint8_t) 1) == 1)
   {
@@ -231,20 +235,22 @@ bool KT0803K::setChannel(uint16_t channel)
   if ((channel < 1400) || (channel > 2160)) return false;
   //  need to split over 3 registers
   uint16_t ch = channel;
-  //  register 2
-  uint8_t data = readData(0x02) & 0x7F;
-  data |= (channel & 0x01) << 8;
-  if (writeData(0x02, data) == false) return false;
+
+  uint8_t register2 = readData(0x02) & 0x7F;
+  register2 |= (channel & 0x01) << 7;   //  CHSEL[0]
+  if (writeData(0x02, register2) == false) return false;
   ch >>= 1;
-  //  register 0
-  if (writeData(0x00, ch & 0xFF) == false) return false;
-  //  register 1
+
+  uint8_t register0 = ch & 0xFF;  //  CHSEL[8:1]
+  if (writeData(0x00, register0) == false) return false;
+
   ch >>= 8;
-  data = readData(0x01);
-  data &= 0xF8;   //  keep other bits
-  data |= ch & 0x07;
-  return writeData(0x01, data);
+  uint8_t register1 = readData(0x01);
+  register1 &= 0xF8;         //  CHSEL[11:9]
+  register1 |= (ch & 0x07);  //  CHSEL[11:9]
+  return writeData(0x01, register1);
 }
+
 
 uint16_t KT0803K::getChannel()
 {
@@ -252,15 +258,74 @@ uint16_t KT0803K::getChannel()
   channel <<= 8;
   channel |= readData(0x00);
   channel <<= 1;
-  channel |= (readData(0x02) >> 0x07);
+  channel |= (readData(0x02) >> 7);
   return channel;
 }
 
 
 ///////////////////////////////////////////////////////////
 //
-//  OTHER
+//  KT0803K SPECIFIC
 //
+bool KT0803K::setMono()
+{
+  uint8_t register4 = readData(0x04);
+  if (register4 & (1 << 6))
+  {
+    register4 &= ~(1 << 6);
+    return writeData(0x04, register4);
+  }
+  return true;
+}
+
+
+bool KT0803K::setStereo()
+{
+  uint8_t register4 = readData(0x04);
+  if ((register4 & (1 << 6)) == 0)
+  {
+    register4 |= (1 << 6);
+    return writeData(0x04, register4);
+  }
+  return true;
+}
+
+bool KT0803K::isStereo()
+{
+  uint8_t register4 = readData(0x04);
+  return (register4 & (1 << 6));
+}
+
+
+bool KT0803K::setBass(uint8_t bass)  //  0..3 = 0, 5, 11, 17 dB
+{
+  if (bass > 3) return false;
+  uint8_t register4 = readData(0x04);
+  register4 &= ~0x03;  //  mask off bits
+  register4 |= bass;
+  return writeData(0x04, register4);
+}
+
+
+uint8_t KT0803K::getBass()
+{
+  uint8_t register4 = readData(0x04);
+  return register4 & 0x03;
+}
+
+
+bool KT0803K::powerOK()
+{
+  uint8_t register0F = readData(0x0F);
+  return (register0F & (1 << 4)) > 0;
+}
+
+
+bool KT0803K::silenceDetected()
+{
+  uint8_t register0F = readData(0x0F);
+  return (register0F & (1 << 2)) > 0;
+}
 
 
 //  -- END OF FILE --